01-faq-graphrag-demo/requirements.txt
strands-agents[openai]>=1.27.0
neo4j>=5.28.0
neo4j-graphrag[openai]>=1.13.0
faiss-cpu>=1.9.0
sentence-transformers>=3.4.0
python-dotenv>=1.0.1
